+++ internal/controller/eventing/integrationtests/validation/integration_test.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"github.com/kyma-project/eventing-manager/api/v1alpha1"
 	"github.com/kyma-project/eventing-manager/test"
-	"github.com/kyma-project/eventing-manager/test/integration"
 	eventingMatchers "github.com/kyma-project/eventing-manager/test/matchers"
+	"github.com/kyma-project/eventing-manager/test/utils/integration"
 	"github.com/onsi/gomega"
 	gomegatypes "github.com/onsi/gomega/types"
 	"github.com/stretchr/testify/require"
